Functionality: add update

+++ backend/core/api/views.py
@@ -1,8 +1,10 @@
-import common.api_config as APICnf
+import common.api_config as Cnf
 from api.models import Student
+from rest_framework import generics
 from rest_framework.views import APIView
-from rest_framework.permissions import IsAuthenticated
-from api.serializers import StudentSerializer
+from django.contrib.auth.models import User
+from rest_framework.permissions import IsAuthenticated, AllowAny
+from api.serializers import StudentSerializer, UserSerializer, RegisterSerializer
 
 
 class HealthView(APIView):
@@ -12,7 +14,7 @@ class HealthView(APIView):
     permission_classes = (IsAuthenticated, )
   
     def get(self, request, format=None):
-        res = APICnf.prepareSuccessResponse('Running fine!')
+        res = Cnf.prepareSuccessResponse('Running fine!')
         return res
 
 
@@ -22,19 +24,30 @@ class UserView(APIView):
     '''
     permission_classes = (IsAuthenticated, )
   
-    def get(self, request, format=None):        
+    def get(self, request):        
         try:
+            user = User.objects.get(id=request.user.id)
             userInfo = {
-                'uname': request.user.username,
-                'email': request.user.email,
-                'fname': request.user.first_name,
-                'lname': request.user.last_name,
+                'uname': user.username,
+                'email': user.email,
+                'fname': user.first_name,
+                'lname': user.last_name,
+                'admin': user.is_superuser,
+                'id': user.id,
             }
-            res = APICnf.prepareSuccessResponse(userInfo)
+            res = Cnf.prepareSuccessResponse(userInfo)
             return res
         except Exception as e:
-            return APICnf.prepareErrorResponse('default')
-            
+            return Cnf.prepareErrorResponse('default')
+
+
+class RegisterUserView(generics.CreateAPIView):
+    '''
+    Creates user instance in User model
+    '''
+    permission_classes = (AllowAny,)
+    serializer_class = RegisterSerializer
+
 
 class StudentListView(APIView):
     '''
@@ -46,21 +59,26 @@ class StudentListView(APIView):
         try:
             students = Student.objects.all()
             serializedData = StudentSerializer(students, many=True)
-            res = APICnf.prepareSuccessResponse(serializedData.data)
+            res = Cnf.prepareSuccessResponse(serializedData.data)
             return res
         except Exception as e:
-            return APICnf.prepareErrorResponse('dberr_get')
+            print(f"[ERROR]: {e}")
+            return Cnf.prepareErrorResponse('dberr_get')
         
     
     def post(self, request):
         try:
             serializedData = StudentSerializer(data=request.data)
+            res = None
             if serializedData.is_valid():
                 serializedData.save()
-                res = APICnf.prepareSuccessResponse(serializedData.data, 201)
+                res = Cnf.prepareSuccessResponse(serializedData.data, 201)
+            else:
+                res = Cnf.prepareErrorResponse('dberr_post', 500, "Invalid input for the fields!")
             return res
         except Exception as e:
-            return APICnf.prepareErrorResponse('dberr_post')
+            print(f"[ERROR]: {e}")
+            return Cnf.prepareErrorResponse('dberr_post')
 
 
 class StudentDetailView(APIView):
@@ -71,39 +89,54 @@ class StudentDetailView(APIView):
 
     def fetchRecordBasedOnId(self, id):
         try:
-            return Student.objects.get(pk = id)
+            return Student.objects.get(user=id)
         except Exception as e:
+            print(f"[ERROR]: {e}")
             raise Exception(str(e))
 
     
     def get(self, request, id):
         try:
             student = self.fetchRecordBasedOnId(id)
-            serializedData = StudentSerializer(student)
-            res = APICnf.prepareSuccessResponse(serializedData.data)
+            res = None
+            if (student):
+                serializedData = StudentSerializer(student)
+                res = Cnf.prepareSuccessResponse(serializedData.data)
+            else:
+                res = Cnf.prepareErrorResponse('dberr_get', 400)
             return res
         except Exception as e:
-            return APICnf.prepareErrorResponse('dberr_get', 400, str(e))
+            print(f"[ERROR]: {e}")
+            return Cnf.prepareErrorResponse('dberr_get', 400, str(e))
 
 
     def put(self, request, id):
         try:
             student = self.fetchRecordBasedOnId(id)
             serializedData = StudentSerializer(student, data=request.data)
+            res = None
             if serializedData.is_valid():
                 serializedData.save()
-                res = APICnf.prepareSuccessResponse(serializedData.data, 200)
+                res = Cnf.prepareSuccessResponse(serializedData.data, 200)
+            else:
+                res = Cnf.prepareErrorResponse('dberr_put', 500)
             return res
         except Exception as e:
-            return APICnf.prepareErrorResponse('dberr_put', 400, str(e))
+            print(f"[ERROR]: {e}")
+            return Cnf.prepareErrorResponse('dberr_put', 400, str(e))
 
 
     def delete(self, request, id):
         try:
             student = self.fetchRecordBasedOnId(id)
-            serializedData = StudentSerializer(student)
-            student.delete()
-            res = APICnf.prepareSuccessResponse(serializedData.data, 204)
+            res = None
+            if (student):
+                serializedData = StudentSerializer(student)
+                student.delete()
+                res = Cnf.prepareSuccessResponse(serializedData.data, 204)
+            else:
+                res = Cnf.prepareErrorResponse('dberr_del', 500)
             return res
         except Exception as e:
-            return APICnf.prepareErrorResponse('dberr_del', 400, str(e))
+            print(f"[ERROR]: {e}")
+            return Cnf.prepareErrorResponse('dberr_del', 400, str(e))
